RM-1429 (The user with read-only permissions on the item can unfreeze it via deleting hold)
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/modules/recordsmanagement/HEAD@72277 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ rm-server/source/java/org/alfresco/module/org_alfresco_module_rm/hold/HoldServiceImpl.java
@@ -418,6 +418,27 @@ public class HoldServiceImpl extends ServiceBaseImpl
             throw new AlfrescoRuntimeException("Can't delete hold, becuase passed node is not a hold. (hold=" + hold.toString() + ")");
         }
 
+        List<NodeRef> held = getHeld(hold);
+        List<String> heldNames = new ArrayList<String>();
+        for (NodeRef nodeRef : held)
+        {
+            if (permissionService.hasPermission(hold, RMPermissionModel.FILING) == AccessStatus.ALLOWED)
+            {
+                heldNames.add((String) nodeService.getProperty(nodeRef, ContentModel.PROP_NAME));
+            }
+        }
+
+        if (heldNames.size() > 0)
+        {
+            StringBuilder sb = new StringBuilder();
+            for (String name : heldNames)
+            {
+                sb.append("\n ");
+                sb.append(name);
+            }
+            throw new AlfrescoRuntimeException("Can't delete hold, because filing permissions for the following items are needed: " + sb.toString());
+        }
+
         // delete the hold node
         nodeService.deleteNode(hold);
     }
